adiciona hook useAuthGuard para gerenciar autenticação e redirecionamento
