+++ frontend/app/search.tsx
@@ -1,14 +1,15 @@
-import { View, TouchableOpacity, Text, TextInput } from "react-native";
+import { View, TouchableOpacity, Text, TextInput, ScrollView } from "react-native";
 import { Ionicons } from "@expo/vector-icons";
 import React, { useState } from "react";
+import { useAuthGuard } from "@/hooks/useAuthGuard";
 
 export default function Search() {
   const [searchText, setSearchText] = useState("");
-
+  useAuthGuard();
   const handleClear = () => setSearchText("");
 
   return (
-    <View className="flex-1 bg-black justify-start">
+    <ScrollView className="flex-1 bg-black justify-start">
       {/* Barra de pesquisa */}
       <View className="p-4 flex-row items-center w-[100%] mt-32 px-16 rounded-full">
         <TouchableOpacity className="absolute left-4">
@@ -54,6 +55,6 @@ export default function Search() {
 
 
 
-    </View>
+    </ScrollView>
   );
 }
