+++ frontend/hooks/useAuthGuard.tsx
@@ -1,0 +1,30 @@
+import { useEffect } from "react";
+import { useRouter } from "expo-router";
+import AsyncStorage from "@react-native-async-storage/async-storage";
+import Constants from "expo-constants";
+
+export function useAuthGuard() {
+  const router = useRouter();
+
+  useEffect(() => {
+    const checkAuth = async () => {
+      const API_URL = (Constants.expoConfig?.extra as any)?.API_URL as string | undefined;
+      const token = await AsyncStorage.getItem("auth_token");
+      if (!API_URL || !token) {
+        router.replace("/login");
+        return;
+      }
+      try {
+        const res = await fetch(`${API_URL}/auth/me`, {
+          headers: { Authorization: `Bearer ${token}` },
+        });
+        if (!res.ok) {
+          router.replace("/login");
+        }
+      } catch {
+        router.replace("/login");
+      }
+    };
+    checkAuth();
+  }, [router]);
+}
